tools/rocs-cli/pyproject.toml
[project]
name = "rocs-cli"
version = "0.1.4"
description = "Minimal ROCS CLI for ai-society"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pyyaml>=6.0.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
tooling = [
  "yq>=3.4.3",
]

[project.scripts]
rocs = "rocs_cli.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 120
extend-exclude = [
  "artifacts",
  "build",
  "src/rocs_cli.egg-info",
]

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["E501"]

[tool.ty]

[tool.ty.environment]
python-version = "3.11"
